Hello Michael:
What if we start like:

"
3.  Overview

   This specification extends [RFC8505] and inherits from [RFC8928] to
   provide a registration method - called subscription in this case -
   for anycast and multicast address.  [RFC8505] is agnostic to the
   routing protocol in which the address may be redistributed.

   In classical networks, [RFC8505] may be used for an ND proxy
   operation as specified in [RFC8929], or redistributed in a full-
   fledged routing protocol such as EVPN
   [I-D.thubert-bess-secure-evpn-mac-signaling] or RIFT
   [I-D.ietf-rift-rift].  The device mobility can be gracefully
   supported as long has the routers can exchange and make sense of the
   sequence counter in the TID field of the EARO.
"

Works?

To your question on PASA:

A PASA + RPL combo would use RPL MOP 0 since the routing down is signaled in the address so RPL does not play there.
The steps would be MOP 0 defines the DODAG and the default route. Normal 6LoWPAN + RPL for packets up towards the root.
Then the nodes would form their own PASA based on the PASA of their preferred parent and register that address to their applications in the backend.
Then the application would speak directly to those addresses. The root would turn the address in a PASA-6LoRH which handles the routing down.

To Luigi: we're probably missing the step where the parent advertises its PASA address to tell the child to form its own. Maybe a flag in the NA or something saying the target is a PASA address. Unless it's well known?

All the best

Pascal

> -----Original Message-----
> From: Michael Richardson <mcr+ietf@sandelman.ca>
> Sent: jeudi 17 novembre 2022 17:33
> To: Pascal Thubert (pthubert) <pthubert@cisco.com>
> Cc: Esko Dijk <esko.dijk@iotconsultancy.nl>; carles.gomez@upc.edu;
> 6lo@ietf.org; ipv6@ietf.org
> Subject: Re: [IPv6] [6lo] WG Last Call on draft-ietf-6lo-multicast-
> registration-11
>
>
> Pascal Thubert \(pthubert\) <pthubert=40cisco.com@dmarc.ietf.org> wrote:
>     > Makes sense to me. What about:
>
> I'm mostly happy with this, but maybe:
>
>     >    In the case of LLNs, RPL [RFC6550] is the routing protocol of
> choice
>     > and [RFC9010] specifies how the unicast address advertised with
>
> Maybe this could mention other choices somewhere?
> What about in the case of non-LLNs?  Would it work with OSPFv3?
> Would it work for /128 prefixes on un-bridged wifi?
>
> Could PASA make use of this? (I'm genuinely unclear)
>
>
> --
> Michael Richardson <mcr+IETF@sandelman.ca>   . o O ( IPv6 IøT consulting )
>            Sandelman Software Works Inc, Ottawa and Worldwide
>
>
>