Hello Erik
Please see below

> On Thu, Nov 9, 2023 at 2:50 PM IETF Secretariat
> <ietf-secretariat-reply@ietf.org> wrote:
>> IESG state changed:
>>
>> New State: AD Evaluation::Revised I-D Needed
>>
>> (The previous state was AD Evaluation)
>>
>> # Internet AD comments for draft-ietf-6lo-multicast-registration-15
>> CC @ekline
>>
>> * comment syntax:
>>   - https://github.com/mnot/ietf-comments/blob/main/format.md
>>
>> ## Comments
>>
>> ### General
>>
>> * The idnits tools highlights this disclaimer text issue:
>>
>> """
>>    (Using the creation date from RFC4861, updated by this document, for
>>    RFC5378 checks: 2004-07-16)
>>
>> -- The document seems to lack a disclaimer for pre-RFC5378 work, but may
>>    have content which was first submitted before 10 November 2008.  If you
>>    have contacted all the original authors and they are all willing to grant
>>    the BCP78 rights to the IETF Trust, then this is fine, and you can ignore
>>    this comment.  If not, you may need to add the pre-RFC5378 disclaimer.
>>    (See the Legal Provisions document at
>>    https://trustee.ietf.org/license-info for more information.)
>> """
>>
>>   Do we already know if an author of 4861 has granted BCP78 rights?  If not,
>>   should this document use the pre-5378 disclaimer?

No I did not

>>   Separately: I will take an action item to reach out to RFC 4861 authors
>>   to see if any have already dealt with this BCP78 rights grant.

Great! many thanks. I guess this will be generally useful for any future extension.

>> ### S6
>>
>> * "This specification Extends RFC6550 to require that"
>>
>>   This paragraph doesn't have any capitalized normative language.  Should
>>   this be reworded to get "REQUIRED" in here somewhere?

The uppercase text is elsewhere in the doc. This text is more like a summary of what's happening so I'd rather not use uppercase here, which could be duplicate at best and a contradiction at worse.

Related: is it still fashionable / useful to use draft-kuehlewind-update-tag language?
>> ### S6.4, S6.5, S7.1, S7.2
>>
>> * Can you explain why there are 2 bit used here instead of just 1?
>>
>>   Whether an address is multicast or not can be determined by inspecting
>>   the address itself.  We should only need 1 bit to distinguish unicast
>>   from anycast... right?

We really care about 4 values: unicast, anycast, multicast and prefix thought the "prefix" value is coming up as an active WG doc and still reserved here. So the 2 bits will be needed in the end.

Arguably, it's true that the multicast property can be inferred from the FF. But isn't it cleaner to signal it all at the same place and get a single source of truth?

Sorry there's a long history of tricks playing foul against the trickster in the end.
>> * What is supposed to happen if the address is a multicast address but the
>>   flags indicate unicast/anycast?  Or vice versa?

That's really a bug that should only appear in early testings, and I'm not keen on spending too much standard effort on it.

Conversely, if we do not signal multicast, how would the unicast/anycast bit be set? That incongruity would show beyond early testing.

For that bug, I suggest we silently ignore (drop the request) or return a status 12.

Suggestion:

"

                                    If the value of the P-Field is
      not consistent with the Registered Address, e.g., the Registered
      Address is a multicast address (section 2.4 of [RFC4291]) and the
      P-Field indicates a value that is not 1, or the other way around,
      then the message, NS(EARO) or EDAR, MUST be dropped, and the
      receiving node MAY either reply with a status of 12 "Invalid
      Registration" or remain silent.

"
>> ### S10
>>
>> * What is the benefit of including a representation of the uptime?
>>
>>   It seems like the primary comparison to be used is:
>>
>>       (this NSSI != the last NSSI seen) : yes or no
>>
>>   rather than performing uptime calculations.
>>
>>   Is it simpler to drop this and expand the bits available for the two
>>   NSSI fields?  This would then be a more generic "generation number
>>   option."

The nodes are not 100% expected to retain state between boots. It's only a SHOULD. When they do, you are correct. But "When persisting is not possible, then the NSSI is randomly  generated"; in that case, we need the time stamp to  ensure that state  generated from info from the peer before it rebooted is cleaned up.

Proposal to update the text as follows:

before

"

                                                                                                    any state
   that was installed in the sender or based on information from the
   sender before it rebooted needs to be reassessed and reinstalled if
   it is still needed.

"

after

"

   If the boot time of the sender is updated to a newer time, any state that was
   installed in the sender MUST be reassessed and reinstalled if it is missing but still needed.
   The U flag not set in a unicast message from the sender indicates that it has lost all state from this node.
   If the U flag is set, the the Peer NSSI field can be used to assess which changes the sender missed.

"

>> ## Nits
>>
>> ### S3
>>
>> * I found the wording of "there is no concept of duplication" to be a bit
>>   awkward.  Perhaps something like "there is no restriction on duplication".

Interesting: Arguably the same address is used by all so it is not duplicated. And if 2 distinct multicast groups were to use the same address that would be potentially immense trouble.

So in fact some duplication may exist and be very wrong. bottom line, we are both inexact here. Let me just drop altogether "there is no concept of duplication with multicast and anycast addresses, and"

>>
>>
>> * "so it generally discouraged" -
>>   "so is generally discouraged"
>>
>> ### S6.1
>>
>> * "allows to hybrid" ->
>>   "allows hybrid use of"
>>
>> ### S7.3
>>
>> * s/FF02::1/ff02::1/g
>>
>>   (per RFC 5952 S4.3)
>>
>> ### S10
>>
>> * "then the NSSI is randomly generated loss"
>>
>>   Does this really mean just "randomly generated"?

no idea where that loss came from!

all applied!

Many thanks Erik :)

Pascal

>>
>>
>> Datatracker URL: https://datatracker.ietf.org/doc/draft-ietf-6lo-multicast-registration/
>>
>>